requirements.txt
django
django-allauth
django-webpack-loader
urllib2
